fix: separate components and library
